reid-system/gorillavision/requirements.txt
torchvision==0.14.1
numpy
opencv-python==4.5.5.64
pandas==1.2.5
pytorch_lightning
scikit-learn
torchmetrics
wandb
kornia
seaborn
tensorflow